+++ example/lib/views/night.dart
@@ -1,6 +1,5 @@
-
-
 import 'package:flutter/material.dart';
+import 'package:flutter_floatwing/flutter_floatwing.dart';
 
 class NightView extends StatefulWidget {
   const NightView({ Key? key }) : super(key: key);
@@ -11,14 +10,23 @@ class NightView extends StatefulWidget {
 
 class _NightViewState extends State<NightView> {
 
-  Color color = Color.fromARGB(255, 213, 222, 54);
+  Color color = Color.fromARGB(255, 192, 200, 41).withOpacity(0.20);
+
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  Window? w;
+
+  var _show = true;
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: MediaQuery.of(context).size.height,
-      width: MediaQuery.of(context).size.width,
-      color: color.withOpacity(0.25),
+      height: _show?MediaQuery.of(context).size.height:0,
+      width: _show?MediaQuery.of(context).size.width:0,
+      color: color,
     );
   }
 }