+++ lib/flutter_floatwing.dart
@@ -1,3 +1,6 @@
-export 'src/window.dart';
 export 'src/plugin.dart';
+export 'src/window.dart';
 export 'src/provider.dart';
+export 'src/event.dart';
+export 'src/utils.dart';
+export 'src/constants.dart';