+++ lib/src/plugin.dart
@@ -1,33 +1,47 @@
 import 'dart:async';
+import 'dart:developer';
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_floatwing/flutter_floatwing.dart';
+import 'package:flutter_floatwing/src/utils.dart';
 import 'package:flutter_floatwing/src/window.dart';
 
 class FloatwingPlugin {
   FloatwingPlugin._() {
     WidgetsFlutterBinding.ensureInitialized();
 
-    _bgChannel.setMethodCallHandler((call) {
-      var id = call.arguments as String;
-      // if we are window egine, should call main engine
-      FloatwingPlugin().windows[id]?.eventManager?.sink(call.method, call.arguments);
-      switch (call.method) {
+    // make sure this only be called once
+    // what happens when multiple window instances
+    // are created and register event handlers?
+    // Window().on(): id -> [Window, Window]
+    // _eventManager = EventManager(_msgChannel);
 
-      }
-      return Future.value(null);
-    });
+    // _bgChannel.setMethodCallHandler((call) {
+    //   var id = call.arguments as String;
+    //   // if we are window egine, should call main engine
+    //   FloatwingPlugin().windows[id]?.eventManager?.sink(call.method, call.arguments);
+    //   switch (call.method) {
+
+    //   }
+    //   return Future.value(null);
+    // });
   }
 
   static const String channelID = "im.zoe.labs/flutter_floatwing";
 
   static final MethodChannel _channel = MethodChannel('$channelID/method');
+
   static final MethodChannel _bgChannel = MethodChannel('$channelID/bg_method');
+
   static final BasicMessageChannel _msgChannel =
       BasicMessageChannel('$channelID/bg_message', JSONMessageCodec());
 
   static final FloatwingPlugin _instance = FloatwingPlugin._();
+
+  /// event manager
+  // EventManager? _eventManager;
 
   /// flag for inited
   bool _inited = false;
@@ -55,10 +69,10 @@ class FloatwingPlugin {
   /// return current window for window's engine
   Window? get currentWindow => _window;
 
-  /// update the current window
-  saveWindow(Window w) {
-    _window = w;
-  }
+  /// i'm window engine, default is the main engine
+  /// if we sync success, we set to true.
+  bool get isWindow => _isWindow;
+  bool _isWindow = false;
 
   factory FloatwingPlugin() {
     return _instance;
@@ -68,20 +82,34 @@ class FloatwingPlugin {
     return _instance;
   }
 
+  /// sync make the plugin to sync windows from services
+  Future<bool> syncWindows() async {
+    var _ws = await _channel.invokeListMethod("plugin.sync_windows");
+    _ws?.forEach((e) {
+      var w = Window.fromMap(e);
+      _windows[w.id] = w;
+    });
+    return true;
+  }
+
   Future<bool> initialize() async {
     if (_inited) return false;
     _inited = true;
 
     // get the callback id
-    final CallbackHandle _cbId = PluginUtilities.getCallbackHandle(_callback)!;
+    // final CallbackHandle _cbId = PluginUtilities.getCallbackHandle(_callback)!;
     // if service started will return all windows
     var map = await _channel.invokeMapMethod("plugin.initialize", {
       // "start_service": true,
-      "callback": _cbId.toRawHandle(),
+      // "callback": _callback,
+
+      // DEPRECATED: use system
       "pixelRadio": window.devicePixelRatio,
+
+      "system": SystemConfig().toMap(),
     });
 
-    print("[plugin] initialize result: $map");
+    log("[plugin] initialize result: $map");
 
     _serviceRunning = map?["service_running"];
     _permissionGranted = map?["permission_grated"];
@@ -92,7 +120,7 @@ class FloatwingPlugin {
       _windows[w.id] = w;
     });
 
-    print("[plugin] there are ${_windows.length} windows already started");
+    log("[plugin] there are ${_windows.length} windows already started");
 
     return true;
   }
@@ -113,27 +141,57 @@ class FloatwingPlugin {
     return await _channel.invokeMethod("plugin.start_service");
   }
 
-  // create window object
-  Future<Window> createWindow(
+  Future<bool> cleanCache() async {
+    return await _channel.invokeMethod("plugin.clean_cache");
+  }
+
+  /// create window to create a window
+  Future<Window?> createWindow(
     String? id,
     WindowConfig config, {
     bool start = false, // start immediately if true
     Window? window,
   }) async {
+    var w = isWindow
+        ? await currentWindow?.createChildWindow(id, config,
+            start: start, window: window)
+        : await internalCreateWindow(id, config,
+            start: start, window: window, channel: _channel);
+    if (w == null) return null;
+    // store current window for window engine
+    // for window engine use, update the current window
+    // if we use create_window first?
+    // _window = w; // we should don't use create_window first!!!
+    // store the window to cache
+    _windows[w.id] = w;
+    return w;
+  }
+
+  // create window object for main engine
+  Future<Window?> internalCreateWindow(
+    String? id,
+    WindowConfig config, {
+    bool start = false, // start immediately if true
+    Window? window,
+    required MethodChannel channel,
+    String name = "plugin.create_window",
+  }) async {
     // check permission first
     if (!await checkPermission()) {
       throw Exception("no permission to create window");
     }
-    var updates =
-        await _channel.invokeMapMethod<String, dynamic>("plugin.create_window", {
+
+    // store the window first
+    // window.id can't be updated
+    // for main engine use
+    // if (window != null) _windows[window.id] = window;
+    var updates = await channel.invokeMapMethod(name, {
       "id": id,
       "config": config.toMap(),
       "start": start,
     });
-    var w = (window ?? Window()).applyMap(updates);
-    // store the window to cache
-    _windows[w.id] = w;
-    return w;
+    // if window is not created, new one
+    return updates == null ? null : (window ?? Window()).applyMap(updates);
   }
 
   /// ensure window make sure the window object sync from android
@@ -141,60 +199,29 @@ class FloatwingPlugin {
   /// you should only call this in the window engine
   /// if only main as entry point, it's ok to call this
   /// and return nothing
+  // only window engine call this
+  // make sure window engine return only one window from every where
   Future<Window?> ensureWindow() async {
-    return Window().sync();
+    // window object don't have sync method, we must do at here
+    // assert if you are in main engine should call this
+    var map = await Window.sync();
+    log("[window] sync window object from android: $map");
+    if (map == null) return null;
+    // store current window if needed
+    // use the static window first
+    // so sync will return only one instance of window
+    // improve this logic
+    // means first time call sync, just create a new window
+    if (_window == null) _window = Window();
+    _window!.applyMap(map);
+    _isWindow = true;
+    return _window;
   }
 
-  // Future<bool> startWindow(String id) async {
-  //   return await _bgChannel.invokeMethod("service.start_window", {
-  //     "id": id
-  //   });
-  // }
-
-  // Future<bool> closeWindow(String id, {bool force = false}) async {
-  //   return await _bgChannel.invokeMethod("service.close_window", {
-  //     "id": id,
-  //     "force": force,
-  //   });
-  // }
-
-  // Future<Map<dynamic, dynamic>?> updateWindow(
-  //     String id, WindowConfig config) async {
-  //   var updates = await _bgChannel.invokeMapMethod("service.update_window", {
-  //     "id": id,
-  //     "config": config.toMap(),
-  //   });
-  //   // store the window to cache
-  //   _windows[id]?.applyMap(updates);
-  //   return updates;
-  // }
-
-  // Future<bool> showWindow(String id, bool v) async {
-  //   return await _bgChannel.invokeMethod("service.show_window", {
-  //     "id": id,
-  //     "visible": v
-  //   });
-  // }
-
-  // Window? getWindow(String id) {
-  //   return _windows[id];
-  // }
-
-  /// sync window return a window from engine
-  /// when in the flutter don't known window object.
-  // Future<Window?> syncWindow() async {
-  //   if (_window != null) return _window;
-  //   // create window and sync from service
-  //   return Window(null).sync();
-
-  //   // var map = await MethodChannel('$channelID/bg_method/window')
-  //   //   .invokeMapMethod("window.init");
-  //   // print("receive init call from android: $map");
-  //   // var w = Window.fromMap(map);
-  //   // // set to static
-  //   // FloatwingPlugin()._window = w;
-  //   // return w;
-  // }
-
-  static void _callback() async {}
+  /// `on` register event handlers for all windows
+  /// or we can use stream mode
+  FloatwingPlugin on(EventType type, WindowListener callback) {
+    // TODO:
+    return this;
+  }
 }